pyproject.toml
[project]
name = "bags-sniper"
version = "0.1.0"
description = "Bags.fm AI Sniping Bot - Deployer Intelligence Trading System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "solana>=0.34.0",
    "solders>=0.21.0",
    "httpx>=0.27.0",
    "asyncio>=3.4.3",
    "python-telegram-bot>=21.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "redis>=5.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "base58>=2.1.0",
]

[project.scripts]
bags-sniper = "bags_sniper.cli:main"
bags-sniper-bot = "bags_sniper.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.2.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bags_sniper"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
